Clean up argument parsing and DB_PATH usage
Sorry to my future self for the big ball of change here.  I was
experimenting and the CLI stuff was not straightforward.  But
the final solution was quite satisfying.

I also included removal of example functionality from my
starter app (the mini-grep stuff).  I just wanted this saved
without trying to clean up and separate what was in the
workspace.  Do better next time.